Put more helper functions in epyccel_utilities.py
