+++ tests/epyccel/test_multiple_results.py
@@ -1,14 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
-from pyccel import epyccel
+from epyccel_utilities import compare_epyccel
+
 from pyccel.decorators import pure
-
-
-# ==============================================================================
-def compare_epyccel(f1, *args, language):
-    f2 = epyccel(f1, language=language)
-    out1 = f1(*args)
-    out2 = f2(*args)
-    assert all(r1 == r2 for r1, r2 in zip(out1, out2))
 
 
 # ==============================================================================
@@ -18,7 +11,7 @@ def test_const_int_float(language):
     def const_int_float():
         return 1, 3.4
 
-    compare_epyccel(const_int_float, language=language)
+    compare_epyccel(const_int_float, language)
 
 
 # ...
@@ -28,7 +21,7 @@ def test_const_complex_bool_int(language):
     def const_complex_bool_int():
         return 1 + 2j, False, 8
 
-    compare_epyccel(const_complex_bool_int, language=language)
+    compare_epyccel(const_complex_bool_int, language)
 
 
 # ...
@@ -38,4 +31,4 @@ def test_expr_float_int_bool(language):
     def expr_complex_int_bool(n: "int"):
         return 0.5 + n * 1j, 2 * n, n == 3
 
-    compare_epyccel(expr_complex_int_bool, 3, language=language)
+    compare_epyccel(expr_complex_int_bool, language, 3)
